Discover() returns immediately with a channel
If you don't read the channel, the process might just end. Consume the channel
until it is closed (when discovery times out).  Discover still also records
bulbs in the light collection.

This is still only finding bulbs when I use another client on the network. It's
as if the bulbs are ignoring my GetLight message, so I suspect it's not quite
structured right, yet. Perhaps the site/target/tagged combo needs to be
directly addressed.