Only send a single LightGet during discovery
This lessens the impact of GH issue #14 by only requesting a single LightState
from each bulb. In some slow/non-reliable networks this may mean we miss some light
state, but I'm OK with that for now.

If/when we move to a "constant discovery" process, it might make sense to
regularly solicit light states, though.